% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nasty}
\alias{nasty}
\title{Example data for pkgproject package}
\format{
A data frame with 970 rows and 13 variables:
\describe{
  \item{Date}{Date of sample}
  \item{Operator}{Name of operator}
  \item{Process.Sample}{Sample number within the request}
  \item{Product size}{Size of the measured product}
  \item{Measure 1 to 6}{6 columns for 6 measurments taken within one sample}
  \item{Target.Value}{Value of reference for measured objects}
  \item{Tare}{Weight of cantainer}
  \item{Request}{Request number}
}
}
\usage{
nasty
}
\description{
A dataset containing simulated measurements of a business process
with other attributes.
}
\keyword{datasets}
