% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_demo.R
\name{Run_demo}
\alias{Run_demo}
\title{Run Shiny app for estimate_area function}
\usage{
Run_demo()
}
\value{
A \code{Shiny app} containing the following elements:
 \describe{
     \item{sidebarPanel}{Sidebar where one can input the number of points to simulate and the seed}
     \item{mainPanel}{A plot of the area, as well as textual output for the value of the area and the time of execution}
}
}
\description{
Runs an interactive Shiny app that allows the user
 adjust the inputs of the estimate_area() function.
}
\author{
Özgür Aydemir, Sophie La Gennusa, Louis del Perugia, Daniel Szenes, Francesca Darino
}
